# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Task Manager API",
        "version": "1.0.0"
    },
    "paths": {
        "/api/tasks": {
            "get": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Get all tasks",
                "operationId": "0ff3f1bde953dec2e0994ecc5b08316e",
                "responses": {
                    "200": {
                        "description": "List of tasks",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer"
                                            },
                                            "title": {
                                                "type": "string"
                                            },
                                            "description": {
                                                "type": "string"
                                            },
                                            "completed": {
                                                "type": "boolean"
                                            },
                                            "due_date": {
                                                "type": "string",
                                                "format": "date"
                                            },
                                            "user_id": {
                                                "type": "integer"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            },
            "post": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Create a new task",
                "operationId": "ae58b66574c44f2e666ec839a1044a13",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "user_id"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Buy groceries"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Milk, bread, cheese"
                                    },
                                    "due_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-07-15"
                                    },
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Task created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "title": {
                                            "type": "string"
                                        },
                                        "description": {
                                            "type": "string"
                                        },
                                        "due_date": {
                                            "type": "string",
                                            "format": "date"
                                        },
                                        "completed": {
                                            "type": "boolean"
                                        },
                                        "user_id": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "/api/external": {
            "get": {
                "tags": [
                    "External"
                ],
                "summary": "Get posts from external API",
                "operationId": "4e5534bb430ee656dedcc2160cac7a7a",
                "responses": {
                    "200": {
                        "description": "List of posts",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "/api/tasks/complete/{id}": {
            "put": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Mark a task as complete",
                "operationId": "6c5ac0be218da79e5bcbe71573799bdc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the task to mark complete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Task marked as complete",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "completed": {
                                            "type": "boolean"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "/api/notify-overdue": {
            "get": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Send notifications for overdue tasks",
                "operationId": "72dfb85a8b6eca44d32d5eb231fe0f7c",
                "responses": {
                    "200": {
                        "description": "Notifications sent successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "Notificaciones enviadas"
                                        },
                                        "cantidad": {
                                            "type": "integer",
                                            "example": 3
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "/api/tasks/{id}": {
            "get": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Get a task by its ID",
                "operationId": "176bd56869dd444be0a693aedfc1ba53",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the task",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Task details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "title": {
                                            "type": "string"
                                        },
                                        "description": {
                                            "type": "string"
                                        },
                                        "completed": {
                                            "type": "boolean"
                                        },
                                        "due_date": {
                                            "type": "string",
                                            "format": "date"
                                        },
                                        "user_id": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Tasks",
            "description": "Tasks"
        },
        {
            "name": "External",
            "description": "External"
        },
        {
            "name": "Notifications",
            "description": "Notifications"
        }
    ]
}